change type of column

+++ src/main/kotlin/com/example/kotlin_rest/Dto/CountyDto.kt
@@ -1,9 +1,9 @@
 package com.example.kotlin_rest.Dto
 
-data class CountyDto (
+data class CountyDto(
 
     val id: Int,
     val name: String,
-    val population: Long,
+    val population: Int,
 
     )
